requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.5.0
propcache==0.3.2
pymongo==4.13.2
python-dotenv==1.1.0
python-telegram-bot==20.7
pytz==2025.2
requests==2.32.4
sniffio==1.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
gunicorn